Add control and tools directories to MANIFEST.in, update move function in index.html, and add gray_to_rgb function in image_processing.py Set up standards for microscope control functions.
