+++ MANIFEST.in
@@ -1,2 +1,4 @@
 include squid_control/icon/*
 include squid_control/images/*
+include squid_control/control/*
+include squid_control/tools/*
