+++ docs/index.html
@@ -379,7 +379,7 @@
                 //     moveZ = zMove * multiplier;
                 // }
 
-                window.microscopeControl.move({x: moveX, y: moveY, z: moveZ, _rkwargs: true});
+                window.microscopeControl.move_by_distance({x: moveX, y: moveY, z: moveZ, _rkwargs: true});
             }
 
         };
