+++ squid_control/__init__.py
@@ -1,4 +1,4 @@
-from .control.core_reef import *
-from .control.camera import *
+from .control import *
+from .control import *
 
 __version__ = "1.0.0"
